// server/src/service/admin/admin.interface.ts
export interface IPageSearchAdminResult {
  id: number;
  name: string;
  nickName: string;
  phone: string;
  status: number;
  sex: number;
  level: number;
}

export interface IPageSearchGamerResult {
  id: number;
  name: string;
  username: string;
  phone: string;
  status: number;
  sex: number;
  company: string;
}